Add README.md and small refactorings
