+++ ij-core/src/main/kotlin/org/jetbrains/research/ij/server/analysis/Analyzer.kt
@@ -1,7 +1,8 @@
 package org.jetbrains.research.ij.server.analysis
 
 import com.intellij.psi.PsiElement
+import java.io.Serializable
 
 interface Analyzer {
-    fun analyze(psiElement: PsiElement): Any?
+    fun analyze(psiElement: PsiElement): Serializable
 }
