+++ ij-server/src/main/resources/data/watch/task.json
@@ -1,6 +1,7 @@
 {
-  "inputPath": "/Users/maria.tigina/IdeaProjects/IJServer/ij-server/src/main/resources/data/input",
-  "outputPath": "/Users/maria.tigina/IdeaProjects/IJServer/ij-server/src/main/resources/data/output/output.txt",
-  "jarPath": "/Users/maria.tigina/IdeaProjects/IJServer/ij-server/src/main/resources/data/jar",
+  "inputType": "FILE",
+  "inputPath": "{PATH_TO_PROJECT_DIR}/IJServer/ij-server/src/main/resources/data/input",
+  "outputPath": "{PATH_TO_PROJECT_DIR}/IJServer/ij-server/src/main/resources/data/output/output.txt",
+  "jarPath": "{PATH_TO_PROJECT_DIR}/IJServer/ij-server/src/main/resources/data/jar",
   "analyzerClassName": "org.jetbrains.research.ij.server.analysis.CallExpressionsAnalyzer"
 }
